CheckSjisChar 関数の第2引数の型を int から size_t に変更 x64 ビルドで警告が出る数を減らす為の変更 関数の呼び出し元では pr_end-pr という記述でバッファ長を求めており、 pr_end の値の作り方から考えて演算結果が負の値になる事は無いと思われる

+++ sakura_core/charset/codechecker.cpp
@@ -141,7 +141,7 @@ const char* TABLE_JISESCDATA[] = {
 
 	@return 確認した文字の長さ
 */
-int CheckSjisChar( const char* pS, const int nLen, ECharSet *peCharset )
+int CheckSjisChar( const char* pS, const size_t nLen, ECharSet *peCharset )
 {
 	unsigned char uc;
